[BUG] fixes keyboard shortcuts context
This is adding the `when` option to the keyboard shortcuts, so that they trigger only within the editor, and not for example in the terminal.

Before you couldn't clear a terminal window with `CMD+K` anymore as it was triggering this extension's shortcuts. Now the shortcuts are limited to run only when an editor has the focus.